Refactor operations (#644)
* move stuff from network_operations.py to linalg.py

* move stuff

* move tests to linalg.py

* add tests

* add imports

* formatting

* formatting

* formatting

* remove tensornetwork import

* fix tests

* fix tests
