+++ tensornetwork/__init__.py
@@ -1,9 +1,61 @@
+
+from tensornetwork.network_components import (
+    BaseNode,    
+    CopyNode,
+    Edge,    
+    Node,
+    NodeCollection
+    )
+from tensornetwork.network_operations import (
+    check_connected,
+    check_correct,
+    contract_trace_edges,
+    copy,
+    get_all_edges,
+    get_all_nodes,
+    get_neighbors,
+    get_subgraph_dangling,
+    reachable,
+    reduced_density,
+    remove_node,
+    replicate_nodes,
+    split_node,
+    split_node_full_svd,
+    split_node_qr,
+    split_node_rq,
+    switch_backend
+)
+
 #pylint: disable=line-too-long
-from tensornetwork.network_components import Node, Edge, CopyNode, BaseNode, NodeCollection
-#pylint: disable=line-too-long
-from tensornetwork.network_operations import norm, conj, copy, replicate_nodes, transpose, split_node, split_node_qr, split_node_rq, split_node_full_svd, reachable, check_connected, check_correct, get_all_nodes, get_all_edges, remove_node, contract_trace_edges, get_subgraph_dangling, reduced_density, switch_backend, get_neighbors, kron
-#pylint: disable=line-too-long
-from tensornetwork.network_components import contract, contract_copy_node, contract_between, outer_product, outer_product_final_nodes, contract_parallel, flatten_edges, split_edge, slice_edge, get_all_nondangling, get_all_dangling, flatten_all_edges, flatten_edges_between, get_parallel_edges, get_shared_edges
+from tensornetwork.linalg.linalg import (
+    conj,    
+    eye,
+    kron,
+    norm,
+    ones,
+    randn,
+    random_uniform,
+    transpose,
+    zeros,
+    )
+
+from tensornetwork.network_components import (
+    contract,
+    contract_between,
+    contract_copy_node,
+    contract_parallel,
+    flatten_all_edges,    
+    flatten_edges,
+    flatten_edges_between,
+    get_all_nondangling,
+    get_all_dangling,
+    get_parallel_edges,
+    get_shared_edges,
+    outer_product,        
+    outer_product_final_nodes,
+    slice_edge,    
+    split_edge
+    )
 from tensornetwork.backends.base_backend import BaseBackend
 from tensornetwork.network_components import connect, disconnect
 from tensornetwork.ncon_interface import ncon, ncon_network
