+++ tensornetwork/network_operations.py
@@ -27,89 +27,6 @@ from tensornetwork.network_components import connect, contract_parallel
 Tensor = Any
 
 
-def norm(node: BaseNode) -> Tensor:
-  """The L2 norm of `node`
-
-  Args:
-    node: A `BaseNode`. 
-
-  Returns:
-    The L2 norm.
-
-  Raises:
-    AttributeError: If `node` has no `backend` attribute.
-  """
-  if not hasattr(node, 'backend'):
-    raise AttributeError('Node {} of type {} has no `backend`'.format(
-        node, type(node)))
-  return node.backend.norm(node.tensor)
-
-
-def conj(node: BaseNode,
-         name: Optional[Text] = None,
-         axis_names: Optional[List[Text]] = None) -> BaseNode:
-  """Conjugate a `node`.
-
-  Args:
-    node: A `BaseNode`.
-    name: Optional name to give the new node.
-    axis_names: Optional list of names for the axis.
-
-  Returns:
-    A new node. The complex conjugate of `node`.
-
-  Raises:
-    AttributeError: If `node` has no `backend` attribute.
-  """
-  if not hasattr(node, 'backend'):
-    raise AttributeError('Node {} of type {} has no `backend`'.format(
-        node, type(node)))
-  backend = node.backend
-  if not axis_names:
-    axis_names = node.axis_names
-
-  return Node(
-      backend.conj(node.tensor),
-      name=name,
-      axis_names=axis_names,
-      backend=backend)
-
-
-def transpose(node: BaseNode,
-              permutation: Sequence[Union[Text, int]],
-              name: Optional[Text] = None,
-              axis_names: Optional[List[Text]] = None) -> BaseNode:
-  """Transpose `node`
-
-  Args:
-    node: A `BaseNode`.
-    permutation: A list of int or str. The permutation of the axis.
-    name: Optional name to give the new node.
-    axis_names: Optional list of names for the axis.
-
-  Returns:
-    A new node. The transpose of `node`.
-
-  Raises:
-    AttributeError: If `node` has no `backend` attribute, or if
-      `node` has no tensor.
-    ValueError: If either `permutation` is not the same as expected or
-      if you try to permute with a trace edge.
-  """
-
-  if not hasattr(node, 'backend'):
-    raise AttributeError('Node {} of type {} has no `backend`'.format(
-        node, type(node)))
-
-  perm = [node.get_axis_number(p) for p in permutation]
-  if not axis_names:
-    axis_names = node.axis_names
-
-  new_node = Node(
-      node.tensor, name=name, axis_names=node.axis_names, backend=node.backend)
-  return new_node.reorder_axes(perm)
-
-
 def copy(nodes: Iterable[BaseNode],
          conjugate: bool = False) -> Tuple[dict, dict]:
   """Copy the given nodes and their edges.
@@ -875,6 +792,7 @@ def switch_backend(nodes: Iterable[BaseNode], new_backend: Text) -> None:
     node.tensor = backend.convert_to_tensor(node.tensor)
     node.backend = backend
 
+
 def get_neighbors(node: BaseNode) -> List[Node]:
   """Get all of the neighbors that are directly connected to the given node.
 
@@ -899,40 +817,3 @@ def get_neighbors(node: BaseNode) -> List[Node]:
         neighbors.append(edge.node1)
         neighbors_set.add(edge.node1)
   return neighbors
-
-def kron(nodes: Sequence[BaseNode]) -> BaseNode:
-  """Kronecker product of the given nodes.
-
-  Kronecker products of nodes is the same as the outer product, but the order
-  of the axes is different. The first half of edges of all of the nodes will
-  appear first half of edges in the resulting node, and the second half ot the
-  edges in each node will be in the second half of the resulting node.
-
-  For example, if I had two nodes  :math:`X_{ab}`,  :math:`Y_{cdef}`, and 
-  :math:`Z_{gh}`, then the resulting node would have the edges ordered 
-  :math:`R_{acdgbefh}`.
-   
-  The kronecker product is designed such that the kron of many operators is
-  itself an operator. 
-
-  Args:
-    nodes: A sequence of `BaseNode` objects.
-
-  Returns:
-    A `Node` that is the kronecker product of the given inputs. The first
-    half of the edges of this node would represent the "input" edges of the
-    operator and the last half of edges are the "output" edges of the
-    operator.
-  """
-  input_edges = []
-  output_edges = []
-  for node in nodes:
-    order = len(node.shape)
-    if order % 2 != 0:
-      raise ValueError(
-          f"All operator tensors must have an even order. "
-          f"Found tensor with order {order}")
-    input_edges += node.edges[:order//2]
-    output_edges += node.edges[order//2:]
-  result = outer_product_final_nodes(nodes, input_edges + output_edges)
-  return result
